+++ src/frontend/apps/standalone_site/src/features/HomePage/components/HomePage.spec.tsx
@@ -47,6 +47,8 @@ describe('<HomePage />', () => {
         vod_conversion_enabled: true,
         homepage_banner_title: 'banner title',
         homepage_banner_text: 'banner text',
+        meta_description: 'meta description',
+        meta_title: 'meta title',
       },
     });
 
@@ -76,6 +78,8 @@ describe('<HomePage />', () => {
         vod_conversion_enabled: true,
         homepage_banner_title: '',
         homepage_banner_text: '',
+        meta_description: '',
+        meta_title: '',
       },
     });